Update Snakemake module testing
The workflow.current_basedir seems to be working as expected within
the module without having to define it as a variable in the top level
Snakefile. (Using snakemake==7.32.3)

Note that if a workflow requires files outside of the subdirectory,
then the nextstrain build command _must_ run from the top level directory
in order to have access to the files.